file cleanup and moving code
